requirements.txt
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
huggingface-hub==0.26.5
idna==3.10
inflate64==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
multivolumefile==0.2.3
networkx==3.4.2
numpy==2.2.0
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
psutil==6.1.0
py7zr==0.22.0
pybcj==1.0.2
pycryptodomex==3.21.0
PyMuPDF==1.25.0
pyppmd==1.1.0
python-magic==0.4.27
PyYAML==6.0.2
pyzstd==0.16.2
rarfile==4.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
setuptools==75.6.0
sympy==1.13.1
texttable==1.7.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.1.3
sqlalchemy==2.0.36
ffmpeg-python==0.2.0